backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.8
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# Redis for caching and sessions
redis==5.0.1
# aioredis deprecated and merged into redis-py; use redis>=4.2 for asyncio
# (no need to install aioredis separately)

# Data validation and serialization
pydantic==2.4.2
pydantic-settings==2.0.3

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Email functionality
emails==0.6
jinja2==3.1.2

# File handling and documents
python-docx==0.8.11
PyPDF2==3.0.1
pillow==10.1.0

# Background tasks
celery==5.3.4
flower==2.0.1

# WebSocket support
python-socketio==5.10.0

# Data processing and analytics
pandas==2.1.3
numpy==1.25.2

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS (fixed)
# 'python-cors' does not exist — FastAPI already supports CORS via middleware
# If you want explicit dependency:
fastapi-cors==0.0.6

# Date and time utilities
python-dateutil==2.8.2

# JSON Web Tokens
PyJWT==2.8.0

# AI/ML for lead scoring (optional)
scikit-learn==1.3.2
openai==1.3.5

# Export functionality
openpyxl==3.1.2
reportlab==4.0.7

# Phone number validation
phonenumbers==8.13.25

# Rate limiting
slowapi==0.1.9

# Database connection pooling
SQLAlchemy-Utils==0.41.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production WSGI server
gunicorn==21.2.0

# Health checks and monitoring
prometheus-client==0.19.0

# Timezone handling
pytz==2023.3
